Refactor buy/sell configs into ES module
