+++ server/src/trade/buy/buy.config.js
@@ -1,0 +1,5 @@
+import testStrategy from '../strategies/buy/test-strategy';
+
+export default {
+    strategy: testStrategy,
+};
