--- server/src/trade/buy/buy.config.json
@@ -1,3 +1,0 @@
-{
-    "strategy": "test-strategy"
-}
