+++ server/src/trade/sell/sell.config.js
@@ -1,0 +1,5 @@
+import testStrategy from '../strategies/sell/test-strategy';
+
+export default {
+    strategy: testStrategy,
+};
