--- server/src/trade/sell/sell.config.json
@@ -1,3 +1,0 @@
-{
-    "strategy": "test-strategy"
-}
