+++ server/src/trade/sell/sell.js
@@ -8,7 +8,7 @@ import notifier from 'node-notifier';
 import asyncMap from 'async/map';
 import Broker from '../broker/Broker.js';
 import tradeConfig from '../trade.config.json';
-import sellConfig from './sell.config.json';
+import sellConfig from './sell.config.js';
 import { omitBlacklistedSecurities } from '../strategies/shared/common/common-evals.js';
 import { composeEvalFunctions, getSecurityData } from '../trade-utils.js';
 
@@ -27,14 +27,13 @@ async function run() {
 }
 
 async function getSellList() {
-    const sellStrategy = (await import('../strategies/sell/' + sellConfig.strategy + '.js')).default;
     const positions = await getPositions();
     console.log(chalk.cyan(`\n========== Begin Sell Candidate Evaluation - ${moment().format('MMMM Do YYYY, h:mm:ss a')} ==========`));
     return await filterSeries(positions, async (symbol) => {
         const securityData = await getSecurityData(symbol);
         const evalFunctions = composeEvalFunctions([
             omitBlacklistedSecurities,
-            ...sellStrategy,
+            ...sellConfig.strategy,
         ]);
         const failures = await detectSeries(evalFunctions, async (evalFunc) => {
             const result = await evalFunc(securityData);
