+++ test/unit/server/trade/buy/buy.config.test.js
@@ -1,9 +1,12 @@
-import buyConfig from '../../../../../server/src/trade/buy/buy.config.json';
+import buyConfig from '../../../../../server/src/trade/buy/buy.config.js';
 
 describe('Buy Module Config', () => {
     describe('Type Checking', () => {
-        test('strategy is a string', () => {
-            expect(typeof buyConfig.strategy).toBe('string');
+        test('strategy is an array of functions', () => {
+            expect(buyConfig.strategy.constructor.name).toBe('Array');
+            buyConfig.strategy.forEach(evalFunc => {
+                expect(typeof evalFunc).toBe('function');
+            });
         });
     });
 });
