+++ test/unit/server/trade/sell/sell.config.test.js
@@ -1,9 +1,12 @@
-import sellConfig from '../../../../../server/src/trade/sell/sell.config.json';
+import sellConfig from '../../../../../server/src/trade/sell/sell.config.js';
 
 describe('Buy Module Config', () => {
     describe('Type Checking', () => {
         test('strategy is a string', () => {
-            expect(typeof sellConfig.strategy).toBe('string');
+            expect(sellConfig.strategy.constructor.name).toBe('Array');
+            sellConfig.strategy.forEach(evalFunc => {
+                expect(typeof evalFunc).toBe('function');
+            });
         });
     });
 });
